requirements.txt
mkdocs
mkdocs-material
pyyaml
pdfplumber
beautifulsoup4
nltk

# PDF font extraction support
pillow

# deterministic JSON output across platforms
ujson  
